Split deposits handlers code by table
